implemented the user page and made changes in home to display it if it exists.
