+++ src/types/User.ts
@@ -1,7 +1,8 @@
 type User = {
     name: string,
     location?: string,
-    bio?: string 
+    bio?: string,
+    avatar_url: string
 }
 
 export default User
